pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-matters"
version = "0.1.0"
description = "Local-first DJ/producer automation stack."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "k3ss", email = "dev@k3ss.io"}]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "librosa>=0.10",
  "numpy>=1.26",
  "soundfile>=0.12",
  "pydub>=0.25",
  "yt-dlp>=2024.10.7",
  "httpx>=0.27",
  "aiofiles>=23.2",
  "SQLAlchemy>=2.0",
  "demucs>=4.0.0",
  "python-multipart>=0.0.6",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.6",
  "mypy>=1.9"
]

[tool.pytest.ini_options]
pythonpath = ["."]
